moved test to --tests-- folder for thunks
