+++ src/components/Nav/_Nav.scss
@@ -85,6 +85,7 @@ input[type=checkbox]:checked + label {
 
 h3 span {
   font-weight: 400;
+  font-size: 18px;
 }
 
 .nav-login {
